third_party/WebKit/LayoutTests/platform/win/bluetooth/characteristic/notifications/service-is-removed-expected.txt

This is a testharness.js-based test.
PASS Service is removed. Reject with InvalidStateError. 
Harness: the test ran to completion.
